Added group chat functionality
This is going to be a big one, so just bear with me.

script.js: Got the current_users username through requesting it through
socketio from the backend. Also used that username to change the current
users name on messages from "YOU" to the users username. Also added a
joinChat function to allow the users to switch chats based on the index
number of the group chat.

index.html: Removed the commented code. Also added onclick functionality
to the group chat div's which call the "joinChat" function to change
group chats

(get ready for the big one)

routes.py: Implemented a table of active users by adding users to the
ActiveUsers table when they connect and removing them when they
disconnect.
Added groupchats by adding a system where if the client side requests to
switch chats, they have to provide the chat id they want to switch to.
Then the server will switch the room the client is in to the new room.
Also switch the way messages are sent from it being broadcasted, to it
being send to specific rooms.